# api-gateway/docs/swagger.yaml
basePath: /
definitions:
  github_com_Sharefunds_api-gateway_pkg_auth_pb.LoginResponse:
    properties:
      error:
        type: string
      status:
        type: integer
      token:
        type: string
    type: object
  github_com_Sharefunds_api-gateway_pkg_auth_pb.RegisterResponse:
    properties:
      error:
        type: string
      status:
        type: integer
    type: object
  github_com_Sharefunds_api-gateway_pkg_pool-transactions_pb.CreatePoolInvitationResponse:
    properties:
      createdAt:
        $ref: '#/definitions/timestamppb.Timestamp'
      invitationId:
        type: string
      inviteeId:
        type: string
      poolId:
        type: string
      status:
        type: integer
    type: object
  github_com_Sharefunds_api-gateway_pkg_pool-transactions_pb.CreatePoolParticipantResponse:
    properties:
      joinedAt:
        $ref: '#/definitions/timestamppb.Timestamp'
      participantId:
        type: string
      poolId:
        type: string
      role:
        type: string
      status:
        type: integer
    type: object
...
  github_com_Sharefunds_api-gateway_pkg_pool-transactions_pb.DeletePoolResponse:
    properties:
      poolId:
        type: string
      status:
        type: integer
    type: object
  github_com_Sharefunds_api-gateway_pkg_pool-transactions_pb.GalileoTransactions:
    properties:
      ach_transaction_id:
        type: string
      act_id:
        type: string
      act_type:
        type: string
      amt:
        type: string
      auth_ts:
        type: string
      bal_id:
        type: string
      billing_amt:
        type: string
      billing_curr_code:
        type: string
      description:
        type: string
      details:
        type: string
      external_trans_id:
        type: string
      funding_account_prn:
        type: string
      iac_tax:
        type: number
      iva_tax:
        type: number
      local_amt:
        type: string
      local_curr_code:
        type: string
      mcc:
        type: string
      network_id:
        type: string
      original_auth_id:
        type: string
      pmt_ref_no:
        type: string
      post_ts:
        type: string
      prod_id:
        type: string
      settle_amt:
        type: string
      settle_curr_code:
        type: string
      source_id:
        type: string
      spending_account_prn:
        type: string
      trans_code:
        type: string
    type: object
...
  github_com_Sharefunds_api-gateway_pkg_pool-transactions_pb.GetPoolOverviewResponse:
    properties:
      balance:
        type: number
      status:
        type: integer
      transactionCount:
        type: integer
      transactions:
        items:
          $ref: '#/definitions/github_com_Sharefunds_api-gateway_pkg_pool-transactions_pb.GalileoTransactions'
        type: array
    type: object
  github_com_Sharefunds_api-gateway_pkg_pool-transactions_pb.GetPoolResponse:
    properties:
      balanceGoal:
        type: integer
      createdAt:
        type: string
      description:
        type: string
      participants:
        items:
          $ref: '#/definitions/github_com_Sharefunds_api-gateway_pkg_pool-transactions_pb.Participant'
        type: array
      poolId:
        type: string
      poolname:
        type: string
      status:
        type: integer
      transactions:
        items:
          $ref: '#/definitions/github_com_Sharefunds_api-gateway_pkg_pool-transactions_pb.Transaction'
        type: array
    type: object
  github_com_Sharefunds_api-gateway_pkg_pool-transactions_pb.GetUserAccountOverviewResponse:
    properties:
      balance:
        type: number
      status:
        type: integer
      transactionCount:
        type: integer
      transactions:
        items:
          $ref: '#/definitions/github_com_Sharefunds_api-gateway_pkg_pool-transactions_pb.GalileoTransactions'
        type: array
    type: object
  github_com_Sharefunds_api-gateway_pkg_pool-transactions_pb.GetUserBalanceResponse:
    properties:
      balance:
        type: number
      status:
        type: integer
    type: object
  github_com_Sharefunds_api-gateway_pkg_pool-transactions_pb.GetUserPoolsResponse:
    properties:
      pools:
        items:
          $ref: '#/definitions/github_com_Sharefunds_api-gateway_pkg_pool-transactions_pb.Pool'
        type: array
      status:
        type: integer
    type: object
  github_com_Sharefunds_api-gateway_pkg_pool-transactions_pb.Participant:
    properties:
      createdAt:
        type: string
      joinedAt:
        $ref: '#/definitions/timestamppb.Timestamp'
      participantId:
        type: string
      poolId:
        type: string
      role:
        type: string
      status:
        type: integer
      userId:
        type: string
    type: object
  github_com_Sharefunds_api-gateway_pkg_pool-transactions_pb.Pool:
    properties:
      balanceGoal:
        type: integer
      createdAt:
        type: string
      description:
        type: string
      participants:
        items:
          $ref: '#/definitions/github_com_Sharefunds_api-gateway_pkg_pool-transactions_pb.Participant'
        type: array
      poolId:
        type: string
      poolname:
        type: string
      status:
        type: integer
      transactions:
        items:
          $ref: '#/definitions/github_com_Sharefunds_api-gateway_pkg_pool-transactions_pb.Transaction'
        type: array
    type: object
  github_com_Sharefunds_api-gateway_pkg_pool-transactions_pb.RemovePoolParticipantResponse:
    properties:
      participantId:
        type: string
      poolId:
        type: string
      status:
        type: integer
    type: object
  github_com_Sharefunds_api-gateway_pkg_pool-transactions_pb.SendPaymentToUserResponse:
    properties:
      description:
        type: string
      newBalance:
        type: number
      oldBalance:
        type: number
      status:
        type: integer
    type: object
  github_com_Sharefunds_api-gateway_pkg_pool-transactions_pb.Transaction:
    properties:
      amount:
        type: number
      createdAt:
        $ref: '#/definitions/timestamppb.Timestamp'
      description:
        type: string
      fromAccountId:
        type: string
      status:
        type: string
      toAccountId:
        type: string
      transactionId:
        type: string
    type: object
  github_com_Sharefunds_api-gateway_pkg_pool-transactions_pb.UpdatePoolResponse:
    properties:
      balanceGoal:
        type: integer
      createdAt:
        type: string
      description:
        type: string
      poolname:
        type: string
      status:
        type: integer
    type: object
...
  pkg_auth_routes.LoginRequestBody:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  pkg_auth_routes.RegisterRequestBody:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  pkg_pool-transactions_routes.CreateP2PPaymentRequestBody:
    properties:
      Amount:
        type: number
      Description:
        type: string
    type: object
...
  pkg_pool-transactions_routes.CreatePoolInvitationRequestBody:
    properties:
      UserId:
        type: string
    type: object
...
  pkg_pool-transactions_routes.Pool:
    properties:
      balanceGoal:
        type: integer
      description:
        type: string
      poolName:
        type: string
    type: object
  pkg_pool-transactions_routes.RemovePoolParticipantRequestBody:
    properties:
      UserId:
        type: string
    type: object
  pkg_pool-transactions_routes.UpdatePoolRequestBody:
    properties:
      Pool:
        $ref: '#/definitions/pkg_pool-transactions_routes.Pool'
    type: object
...
  timestamppb.Timestamp:
    properties:
      nanos:
        description: |-
          Non-negative fractions of a second at nanosecond resolution. Negative
          second values with fractions must still have non-negative nanos values
          that count forward in time. Must be from 0 to 999,999,999
          inclusive.
        type: integer
      seconds:
        description: |-
          Represents seconds of UTC time since Unix epoch
          1970-01-01T00:00:00Z. Must be from 0001-01-01T00:00:00Z to
          9999-12-31T23:59:59Z inclusive.
        type: integer
    type: object
host: localhost:3000
info:
  contact: {}
  description: This is the client facing API for Sharefunds/Quilt/PoolParty
  title: Sharefunds Client Facing API
  version: 1.0.1
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login a user, validates client token with firebase and returns
        the client token
      parameters:
      - description: User to login
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/pkg_auth_routes.LoginRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Sharefunds_api-gateway_pkg_auth_pb.LoginResponse'
      summary: Login a user
      tags:
      - Authentication Service
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user, creates an account in Firebase and a user
        in the database
      parameters:
      - description: User to register
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/pkg_auth_routes.RegisterRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Sharefunds_api-gateway_pkg_auth_pb.RegisterResponse'
      summary: Register a new user
      tags:
      - Authentication Service
  /pool-transactions/pools/{PoolID}:
    delete:
      consumes:
      - application/json
      description: Sets Pool status to Inactive, Deactivates secondary account in
        galileo.
      operationId: delete_pool
      parameters:
      - description: Pool ID
        in: path
        name: PoolID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Sharefunds_api-gateway_pkg_pool-transactions_pb.DeletePoolResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete Pool, deletes account in galileo.
      tags:
      - Pool & Transactions Service - Pools
    get:
      consumes:
      - application/json
      description: Fetches pool data, using the poolId, user must be a participant
        in the pool.
      operationId: get_pool_data
      parameters:
      - description: Pool ID
        in: path
        name: PoolID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Sharefunds_api-gateway_pkg_pool-transactions_pb.GetPoolResponse'
      security:
      - ApiKeyAuth: []
      summary: Fetch pool data, including participants and transactions.
      tags:
      - Pool & Transactions Service - Pools
    put:
      consumes:
      - application/json
      description: Updates pool information(name and/or description and/or goal).
      operationId: update_pool
      parameters:
      - description: Pool ID
        in: path
        name: PoolID
        required: true
        type: string
      - description: Pool to update
        in: body
        name: Pool
        required: true
        schema:
          $ref: '#/definitions/pkg_pool-transactions_routes.UpdatePoolRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Sharefunds_api-gateway_pkg_pool-transactions_pb.UpdatePoolResponse'
      security:
      - ApiKeyAuth: []
      summary: Updates pool information.
      tags:
      - Pool & Transactions Service - Pools
  /pool-transactions/pools/{PoolID}/accept-invite:
    post:
      consumes:
      - application/json
      description: Accepts pool invitation, using the poolId.
      operationId: create_pool_participant
      parameters:
      - description: Pool ID
        in: path
        name: PoolID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Sharefunds_api-gateway_pkg_pool-transactions_pb.CreatePoolParticipantResponse'
      security:
      - ApiKeyAuth: []
      summary: Creates ( accepts ) a pool participant , from an invitation.
      tags:
      - Pool & Transactions Service - Pools
  /pool-transactions/pools/{PoolID}/create-invite:
    post:
      consumes:
      - application/json
      description: Creates a pool invitation for a user.
      operationId: create_pool_invitation
      parameters:
      - description: User ID to invite
        in: body
        name: Pool
        required: true
        schema:
          $ref: '#/definitions/pkg_pool-transactions_routes.CreatePoolInvitationRequestBody'
      - description: Pool ID
        in: path
        name: PoolID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Sharefunds_api-gateway_pkg_pool-transactions_pb.CreatePoolInvitationResponse'
      security:
      - ApiKeyAuth: []
      summary: Create Pool Invitation, creates a pool invitation for a user.
      tags:
      - Pool & Transactions Service - Pools
  /pool-transactions/pools/{PoolID}/remove-participant:
    put:
      consumes:
      - application/json
      description: Removes a participant from a pool, using the poolId and participantId,
        user must be the owner of the pool.
      operationId: remove_pool_participant
      parameters:
      - description: User ID to remove from pool
        in: body
        name: Pool
        required: true
        schema:
          $ref: '#/definitions/pkg_pool-transactions_routes.RemovePoolParticipantRequestBody'
      - description: Pool ID
        in: path
        name: PoolID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Sharefunds_api-gateway_pkg_pool-transactions_pb.RemovePoolParticipantResponse'
      security:
      - ApiKeyAuth: []
      summary: Remove a participant from a pool.
      tags:
      - Pool & Transactions Service - Pools
...
  /pool-transactions/pools/list:
    get:
      consumes:
      - application/json
      description: Fetches all pools for a user.
      operationId: get_user_pools
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Sharefunds_api-gateway_pkg_pool-transactions_pb.GetUserPoolsResponse'
      security:
      - ApiKeyAuth: []
      summary: Fetch all pools for a user.
      tags:
      - Pool & Transactions Service
  /pool-transactions/pools/overview/{PoolID}:
    get:
      consumes:
      - application/json
      description: Fetches the overview of a pool, including pool's balance, transactions,
        and transaction count.
      operationId: get_pool_overview
      parameters:
      - description: Pool ID
        in: path
        name: PoolID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Sharefunds_api-gateway_pkg_pool-transactions_pb.GetPoolOverviewResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Pool Overview.
      tags:
      - Pool & Transactions Service - Pools
  /pool-transactions/users/balance/{UserID}:
    get:
      consumes:
      - application/json
      description: Fetches the balance of a user, using the userId.
      operationId: get_user_balance
      parameters:
      - description: User ID
        in: path
        name: UserID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Sharefunds_api-gateway_pkg_pool-transactions_pb.GetUserBalanceResponse'
      security:
      - ApiKeyAuth: []
      summary: Get's user balance.
      tags:
      - Pool & Transactions Service
  /pool-transactions/users/overview/{UserID}:
    get:
      consumes:
      - application/json
      description: Fetches the overview of a user's account, including user's balance,
        transactions, and transaction count.
      operationId: get_user_account_overview
      parameters:
      - description: User ID
        in: path
        name: UserID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Sharefunds_api-gateway_pkg_pool-transactions_pb.GetUserAccountOverviewResponse'
      security:
      - ApiKeyAuth: []
      summary: Get User Account Overview.
      tags:
      - Pool & Transactions Service
  /pool-transactions/users/send-p2p-payment/{ReceiverID}:
    post:
      consumes:
      - application/json
      description: Sends a P2P payment from one user to another, an internal account
        transfer via Galileo.
      operationId: create_p2p_payment
      parameters:
      - description: Payment information
        in: body
        name: PaymentInfo
        required: true
        schema:
          $ref: '#/definitions/pkg_pool-transactions_routes.CreateP2PPaymentRequestBody'
      - description: Receiver ID for payment
        in: path
        name: ReceiverID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Sharefunds_api-gateway_pkg_pool-transactions_pb.SendPaymentToUserResponse'
      security:
      - ApiKeyAuth: []
      summary: Create P2P Payment.
      tags:
      - Pool & Transactions Service
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"